backend/requirements.txt
streamlit>=1.41.0
paddlepaddle==2.6.2
paddleocr>=2.7.0.3
pymupdf>=1.23.8
ollama
pillow
numpy
watchdog